requirements.txt
customtkinter
pillow
python-magic
olefile
pandas
tqdm
